moved prompt bar to left side tab
