+++ components/Chatbar/Chatbar.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useContext, useEffect } from 'react';
+import { useCallback, useContext, useEffect, useState } from 'react';
 
 import { useTranslation } from 'next-i18next';
 
@@ -8,17 +8,28 @@ import { DEFAULT_SYSTEM_PROMPT, DEFAULT_TEMPERATURE } from '@/utils/app/const';
 import { saveConversation, saveConversations } from '@/utils/app/conversation';
 import { saveFolders } from '@/utils/app/folders';
 import { exportData, importData } from '@/utils/app/importExport';
+import { savePrompts } from '@/utils/app/prompts';
+import {
+  CloseSidebarButton,
+  OpenSidebarButton,
+} from '../Sidebar/components/OpenCloseButton';
 
 import { Conversation } from '@/types/chat';
 import { LatestExportFormat, SupportedExportFormats } from '@/types/export';
 import { OpenAIModels } from '@/types/openai';
 import { PluginKey } from '@/types/plugin';
+import { Prompt } from '@/types/prompt';
 
 import HomeContext from '@/pages/api/home/home.context';
 
 import { ChatFolders } from './components/ChatFolders';
 import { ChatbarSettings } from './components/ChatbarSettings';
 import { Conversations } from './components/Conversations';
+
+import { PromptFolders } from './components/PromptFolders';
+import { PromptbarSettings } from './components/PromptbarSettings';
+import { Prompts } from './components/Prompts';
+
 
 import Sidebar from '../Sidebar';
 import ChatbarContext from './Chatbar.context';
@@ -34,12 +45,25 @@ export const Chatbar = () => {
   });
 
   const {
-    state: { conversations, showChatbar, defaultModelId, folders },
+    state: { conversations,
+      showChatbar,
+      defaultModelId,
+      folders,
+      prompts,
+      showPromptbar
+    },
     dispatch: homeDispatch,
     handleCreateFolder,
     handleNewConversation,
     handleUpdateConversation,
   } = useContext(HomeContext);
+
+  enum Tab {
+    CONVERSATIONS = 'CONVERSATIONS',
+    PROMPTS = 'PROMPTS',
+  }
+
+  const [activeTab, setActiveTab] = useState<Tab>(Tab.CONVERSATIONS);
 
   let {
     state: { pluginKeys },
@@ -49,7 +73,7 @@ export const Chatbar = () => {
   }
 
   const {
-    state: { searchTerm, filteredConversations },
+    state: { searchTerm, filteredConversations, promptSearchTerm, filteredPrompts },
     dispatch: chatDispatch,
   } = chatBarContextValue;
 
@@ -193,6 +217,65 @@ export const Chatbar = () => {
     }
   };
 
+  const handleDropPrompt = (e: any) => {
+    if (e.dataTransfer) {
+      const prompt = JSON.parse(e.dataTransfer.getData('prompt'));
+
+      const updatedPrompt = {
+        ...prompt,
+        folderId: e.target.dataset.folderId,
+      };
+
+      handleUpdatePrompt(updatedPrompt);
+
+      e.target.style.background = 'none';
+    }
+  };
+
+  const handleTogglePromptbar = () => {
+    homeDispatch({ field: 'showPromptbar', value: !showPromptbar });
+    localStorage.setItem('showPromptbar', JSON.stringify(!showPromptbar));
+  };
+
+  const handleCreatePrompt = () => {
+    if (defaultModelId) {
+      const newPrompt: Prompt = {
+        id: uuidv4(),
+        name: `Prompt ${prompts.length + 1}`,
+        description: '',
+        content: '',
+        model: OpenAIModels[defaultModelId],
+        folderId: null,
+      };
+
+      const updatedPrompts = [...prompts, newPrompt];
+
+      homeDispatch({ field: 'prompts', value: updatedPrompts });
+
+      savePrompts(updatedPrompts);
+    }
+  };
+
+  const handleDeletePrompt = (prompt: Prompt) => {
+    const updatedPrompts = prompts.filter((p) => p.id !== prompt.id);
+
+    homeDispatch({ field: 'prompts', value: updatedPrompts });
+    savePrompts(updatedPrompts);
+  };
+
+  const handleUpdatePrompt = (prompt: Prompt) => {
+    const updatedPrompts = prompts.map((p) => {
+      if (p.id === prompt.id) {
+        return prompt;
+      }
+
+      return p;
+    });
+    homeDispatch({ field: 'prompts', value: updatedPrompts });
+
+    savePrompts(updatedPrompts);
+  };
+
   useEffect(() => {
     if (searchTerm) {
       chatDispatch({
@@ -213,7 +296,27 @@ export const Chatbar = () => {
     }
   }, [searchTerm, conversations]);
 
-  return (
+
+  useEffect(() => {
+    if (promptSearchTerm) {
+      chatDispatch({
+        field: 'filteredPrompts',
+        value: prompts.filter((prompt) => {
+          const searchable =
+            prompt.name.toLowerCase() +
+            ' ' +
+            prompt.description.toLowerCase() +
+            ' ' +
+            prompt.content.toLowerCase();
+          return searchable.includes(searchTerm.toLowerCase());
+        }),
+      });
+    } else {
+      chatDispatch({ field: 'filteredPrompts', value: prompts });
+    }
+  }, [promptSearchTerm, prompts]);
+
+  return showChatbar ? (
     <ChatbarContext.Provider
       value={{
         ...chatBarContextValue,
@@ -224,25 +327,78 @@ export const Chatbar = () => {
         handlePluginKeyChange,
         handleClearPluginKey,
         handleApiKeyChange,
+        handleCreatePrompt,
+        handleDeletePrompt,
+        handleUpdatePrompt,
       }}
     >
-      <Sidebar<Conversation>
-        side={'left'}
-        isOpen={showChatbar}
-        addItemButtonTitle={t('New chat')}
-        itemComponent={<Conversations conversations={filteredConversations} />}
-        folderComponent={<ChatFolders searchTerm={searchTerm} />}
-        items={filteredConversations}
-        searchTerm={searchTerm}
-        handleSearchTerm={(searchTerm: string) =>
-          chatDispatch({ field: 'searchTerm', value: searchTerm })
-        }
-        toggleOpen={handleToggleChatbar}
-        handleCreateItem={handleNewConversation}
-        handleCreateFolder={() => handleCreateFolder(t('New folder'), 'chat')}
-        handleDrop={handleDrop}
-        footerComponent={<ChatbarSettings />}
-      />
+
+    <div className="flex flex-col h-full">
+        <div className="flex border-b border-gray-200 dark:border-gray-700 mb-5">
+          <button
+            className={`flex-1 p-2 text-sm font-bold ${
+              activeTab === Tab.CONVERSATIONS
+                ? 'border-b-2 border-black dark:border-white'
+                : 'border-b-2 border-transparent'
+            }`}
+            onClick={() => setActiveTab(Tab.CONVERSATIONS)}
+          >
+            {t('Conversations')}
+          </button>
+          <button
+            className={`flex-1 p-2 text-sm font-bold ${
+              activeTab === Tab.PROMPTS
+                ? 'border-b-2 border-black dark:border-white'
+                : 'border-b-2 border-transparent'
+            }`}
+            onClick={() => setActiveTab(Tab.PROMPTS)}
+          >
+            {t('Prompts')}
+          </button>
+        </div>
+        <div className="flex-1 overflow-auto">
+          {activeTab === Tab.CONVERSATIONS && (
+            <Sidebar<Conversation>
+              addItemButtonTitle={t('New chat')}
+              itemComponent={<Conversations conversations={filteredConversations} />}
+              folderComponent={<ChatFolders searchTerm={searchTerm} />}
+              items={filteredConversations}
+              searchTerm={searchTerm}
+              handleSearchTerm={(searchTerm: string) =>
+                chatDispatch({ field: 'searchTerm', value: searchTerm })
+              }
+              handleCreateItem={handleNewConversation}
+              handleCreateFolder={() => handleCreateFolder(t('New folder'), 'chat')}
+              handleDrop={handleDrop}
+            />
+          )}
+          {activeTab === Tab.PROMPTS && (
+          <Sidebar<Prompt>
+          addItemButtonTitle={t('New prompt')}
+          itemComponent={
+            <Prompts
+              prompts={filteredPrompts.filter((prompt) => !prompt.folderId)}
+            />
+          }
+          folderComponent={<PromptFolders />}
+          items={filteredPrompts}
+          searchTerm={promptSearchTerm}
+          handleSearchTerm={(searchTerm: string) =>
+            chatDispatch({ field: 'promptSearchTerm', value: searchTerm })
+          }
+          handleCreateItem={handleCreatePrompt}
+          handleCreateFolder={() => handleCreateFolder(t('New folder'), 'prompt')}
+          handleDrop={handleDropPrompt}
+        />
+          )}
+        </div>
+        <div className="p-2">
+          <ChatbarSettings />
+        </div>
+      </div>
+    <CloseSidebarButton onClick={handleToggleChatbar} side={'left'} />
     </ChatbarContext.Provider>
-  );
+) : (
+    <OpenSidebarButton onClick={handleToggleChatbar} side={'left'} />
+  )
 };
