+++ components/Chatbar/components/PromptbarSettings.tsx
