--- components/Promptbar/Promptbar.state.tsx
@@ -1,11 +1,0 @@
-import { Prompt } from '@/types/prompt';
-
-export interface PromptbarInitialState {
-  searchTerm: string;
-  filteredPrompts: Prompt[];
-}
-
-export const initialState: PromptbarInitialState = {
-  searchTerm: '',
-  filteredPrompts: [],
-};
